+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Add `priority` option for Priority cookie support
   * Fix accidental cookie name/value truncation when given invalid chars
   * Remove quotes from returned quoted cookie value
   * Use `req.socket` over deprecated `req.connection`